+++ integrationTests/consensus/testInitializer.go
@@ -252,11 +252,11 @@ func createConsensusOnlyNode(
 		},
 		RevertAccountStateCalled: func() {
 		},
-		CreateBlockCalled: func(round uint32, haveTime func() bool) (handler data.BodyHandler, e error) {
+		CreateBlockCalled: func(round uint64, haveTime func() bool) (handler data.BodyHandler, e error) {
 			return &dataBlock.Body{}, nil
 		},
-		CreateBlockHeaderCalled: func(body data.BodyHandler, round uint32, haveTime func() bool) (handler data.HeaderHandler, e error) {
-			return &dataBlock.Header{Round: uint32(round)}, nil
+		CreateBlockHeaderCalled: func(body data.BodyHandler, round uint64, haveTime func() bool) (handler data.HeaderHandler, e error) {
+			return &dataBlock.Header{Round: uint64(round)}, nil
 		},
 		MarshalizedDataToBroadcastCalled: func(header data.HeaderHandler, body data.BodyHandler) (bytes map[uint32][]byte, bytes2 map[uint32][][]byte, e error) {
 			mrsData := make(map[uint32][]byte)
@@ -276,7 +276,7 @@ func createConsensusOnlyNode(
 		UnnotarisedBlocksCalled: func() []data.HeaderHandler {
 			return make([]data.HeaderHandler, 0)
 		},
-		SetBlockBroadcastRoundCalled: func(nonce uint64, round int32) {
+		SetBlockBroadcastRoundCalled: func(nonce uint64, round int64) {
 		},
 	}
 	blockChain := createTestBlockChain()
